update load image
1.Немного обновил  структуру  проекта, добавил общую директорию с названием проекта, и в ней уже внутри по директориям детали проекта.
2.Подкоректировал немного файл impred/impred/views.py .
3.Скорректировал файл impred/impred/models.py
4.Добавил заготовку для класса User в impred/user/models.py
По итогу сейчас фото сохраняются и редиректир на страничку prediction.

Далее, можно  брать новые html шаблоны от Андрея и менять их по коду в нашем  проекте для соответсвующих страниц.
Плюс можно разобраться с классом user и реализовать логирование пользователей в приложение.
