+++ impred/impred/models.py
@@ -3,13 +3,16 @@ from django.contrib.auth.models import User
 from PIL import Image
 
 
+# from user.models import СustomUser
+
+
 # Create your models here.
 class NetModels(models.Model):
     name = models.CharField(max_length=30, null=False, unique=True)
     description = models.CharField(max_length=150, null=False)
     created = models.DateTimeField(auto_now_add=True)
     modelpath = models.FileField(upload_to='models/', null=False)
-
+    
     def __str__(self):
         return f"{self.name}"
 
@@ -33,7 +36,7 @@ class Images(models.Model):
     predict = models.ForeignKey(Labels, on_delete=models.CASCADE, null=True)
     created = models.DateTimeField(auto_now_add=True)
     imagepath = models.ImageField(upload_to='images/', null=False)
-    user = models.ForeignKey(User, on_delete=models.CASCADE, default=1)
+    # user = models.ForeignKey(User, on_delete=models.CASCADE, default=1) # временно закоментировал до момента создания класса User
 
     def __str__(self):
         return f"{self.name}"
@@ -47,5 +50,7 @@ class Images(models.Model):
         if img.height > 250 or img.width > 250:
             new_img = (250, 250)
             img.thumbnail(new_img)
-            img.save(self.imagepath)
- 
+            with open(self.imagepath.path, "wb") as f:
+                img.save(f)
+
+            #img.save(self.imagepath.path)
