+++ impred/impred/templates/impred/base.html
