+++ impred/impred/templates/impred/predictimage.html
