+++ impred/user/models.py
@@ -1,0 +1,12 @@
+# Create your models here.
+# from django.contrib.auth.models import AbstractUser
+# from django.db import models
+
+
+# class СustomUser(AbstractUser):
+#     username = models.CharField(max_length=30, null=False)
+#     email = models.EmailField(unique=True)
+#     password = models.CharField(max_length=128)
+
+#     def __str__(self):
+#         return self.username
